# src/openapi.yaml
openapi: "3.0.3"
info:
  title: ScoutOff API
  description: >
    Backend API for ScoutOff — a decentralised football scouting platform
    built on the Stellar blockchain (Soroban smart contracts).
  version: "1.0.0"
  contact:
    name: ScoutOff Engineering
servers:
  - url: /api
    description: Stable API (version-less alias)
  - url: /api/v1
    description: Versioned API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Long-lived scout API key (alternative to a Bearer JWT). Restricted keys
        may only perform operations covered by their granted scopes (#1019).
        Legacy keys without a scope list are unrestricted.
  parameters:
    playerId:
      name: playerId
      in: path
      required: true
      schema:
        type: string
      description: Player's unique identifier (cuid2)
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string

    # ── Player ────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Insufficient role or wallet mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
paths:
...
  /players/{playerId}/trial-offers/{offerId}/accept:
    post:
      tags: [players]
      summary: Accept a trial offer (player owner only)
      operationId: acceptTrialOffer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/playerId"
        - name: offerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Offer accepted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          offerId:
                            type: string
                          playerId:
                            type: string
                          status:
                            type: string
                            enum: [accepted]
                          respondedAt:
                            type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Offer already responded to
  /players/{playerId}/trial-offers/{offerId}/reject:
    post:
      tags: [players]
      summary: Reject a trial offer with optional reason (player owner only)
      operationId: rejectTrialOffer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/playerId"
        - name: offerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Offer rejected
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          offerId:
                            type: string
                          playerId:
                            type: string
                          status:
                            type: string
                            enum: [rejected]
                          reason:
                            type: string
                            nullable: true
                          respondedAt:
                            type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Offer already responded to

  # ── Scouts ────────────────────────────────────────────────────────────────
...